Fonctions raccourcies, réorganisation fichiers. Close #2
