+++ .gitlab-ci.yml
@@ -9,7 +9,7 @@ job:mon_test:
     - sh install.sh
     - python3 ./resolution_coreferences_pronominales/__main__.py
     - for f in resolution_coreferences_pronominales/tests/*.py; do if [ $f != resolution_coreferences_pronominales/tests/test-duree_execution.py ]; then python3 "$f"; fi; done
-    - python3 resolution_coreferences_pronominales/tests/test-duree_execution.py 1 1
+    - python3 resolution_coreferences_pronominales/tests-regression/test-duree_execution.py 1 1
     - sh uninstall.sh
