--- resolution_coreferences_pronominales/data/historique_precision.txt
@@ -1,2 +1,0 @@
-0.9166666666666666
-1
