+++ resolution_coreferences_pronominales/lancer_tests.py
@@ -1,0 +1,4 @@
+from resolution_coreferences_pronominales.tests import test-extra
+
+if __name__ == '__main__':
+
