+++ resolution_coreferences_pronominales/tests-regression/historique_duree_execution.csv
@@ -1,0 +1,5 @@
+cache,nb_reps_fcts,mot.extraction_html(),difference_1,mot.relations_mot(),difference_2,relations_entre_mots(),difference_3,traitements_phrase.informations_pronoms(),difference_4,traitements_phrase.coreferences_phrase(),difference_5,date,heure,mot,liste_mots,phrase
+0.0,5.0,0.61,,13.687000000000001,,29.384,,2.438,,10.502,,26/06/2020,20:25:40,manger,"['chien', 'canaille', 'caniche', 'aboyer', 'être vivant', 'museau']",Le chien est tombé dans le puits. Il s'est cassé le museau. Il va ainsi retenir la leçon.
+1.0,5.0,0.595,,0.017,,0.054000000000000006,,1.0590000000000002,,1.198,,26/06/2020,20:25:54,manger,"['chien', 'canaille', 'caniche', 'aboyer', 'être vivant', 'museau']",Le chien est tombé dans le puits. Il s'est cassé le museau. Il va ainsi retenir la leçon.
+0.0,5.0,0.662,+8.54%,14.346,+4.81%,16.124,-45.13%,2.467,+1.18%,10.454,-0.46%,26/06/2020,20:29:44,manger,"['chien', 'canaille', 'caniche', 'aboyer', 'être vivant', 'museau']",Le chien est tombé dans le puits. Il s'est cassé le museau. Il va ainsi retenir la leçon.
+1.0,5.0,0.605,+1.69%,0.02,+17.69%,0.055,+0.95%,1.066,+0.64%,1.105,-7.77%,26/06/2020,20:29:59,manger,"['chien', 'canaille', 'caniche', 'aboyer', 'être vivant', 'museau']",Le chien est tombé dans le puits. Il s'est cassé le museau. Il va ainsi retenir la leçon.
