+++ resolution_coreferences_pronominales/tests-regression/historique_precision.txt
@@ -1,0 +1,2 @@
+
+0.9166666666666666
