+++ resolution_coreferences_pronominales/tests/test-extraction_mot-duree-d-execution.py
@@ -1,11 +1,12 @@
 import sys
+
 sys.path.append(".")
-from resolution_coreferences_pronominales import extraction_mot
 import time
+from resolution_coreferences_pronominales.coreferences import mot
 
 if __name__ == '__main__':
-    # Paramètres à choisir :
-    mot = 'chien'
+    # Paramètres
+    le_mot = 'chien'
     cache = False
     relation = 'all'
     nb_boucle = 1
@@ -15,11 +16,11 @@ if __name__ == '__main__':
     start_time_total = time.time()
     for i in range(nb_boucle):
         start_time = time.time()
-        relations_mot = extraction_mot.relations_mot(mot, relation, cache)
+        relations_mot = mot.relations_mot(le_mot, relation, cache)
         time1 += time.time() - start_time
 
     # Imprime la durée d'exécution d'extraction_mot.relations_mot (en moyenne)
-    print("relations_mot(" + mot + ", " + relation + ", " + str(cache) + " --- " + str(
+    print("relations_mot(" + le_mot + ", " + relation + ", " + str(cache) + " --- " + str(
         (time1 / nb_boucle)) + " secondes --- par boucle en moyenne")
     print("temps total --- " + str((time.time() - start_time_total)) + " secondes --- pour " + str(
         nb_boucle) + " boucles")
